Move KnownVariantsList to subfolder
